move getCollapseItems to Helper
