+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/LayerCollapse/LayerCollapse.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Collapse, Message } from 'oskari-ui';
 import { Controller, LocaleConsumer } from 'oskari-ui/util';
-import { LayerCollapsePanel } from './LayerCollapsePanel';
 import { Alert } from '../Alert';
 import styled from 'styled-components';
-import { PanelToolContainer } from './PanelToolContainer';
+import { getCollapseItems } from './LayerCollapseHelper';
+
 const StyledCollapse = styled(Collapse)`
     border-radius: 0 !important;
     & > div {
@@ -28,57 +28,17 @@ const StyledCollapse = styled(Collapse)`
     }
 `;
 
-export const getLayerRowModels = (layers = [], selectedLayerIds = [], controller, opts) => {
-    return layers.map(oskariLayer => {
-        return {
-            id: oskariLayer.getId(),
-            model: oskariLayer,
-            selected: selectedLayerIds.includes(oskariLayer.getId()),
-            controller,
-            opts
-        };
-    });
-};
-
 const LayerCollapse = ({ groups, openGroupTitles, selectedLayerIds, opts, controller }) => {
     if (!Array.isArray(groups) || groups.length === 0) {
         return <Alert showIcon type='info' message={<Message messageKey='errors.noResults' />} />;
     }
-
-    const groupItems = groups.map(group => {
-        const layerRows = getLayerRowModels(group.getLayers(), selectedLayerIds, controller, opts);
-        // set group switch active if all layers in group are selected
-        const allLayersOnMap = layerRows.length > 0 && layerRows.every(layer => selectedLayerIds.includes(layer.id));
-        const hasChildren = layerRows.length > 0 || group.getGroups().length > 0;
-        return {
-            key: group.getId(),
-            label: group.getTitle(),
-            className: `t_group gid_${group.getId()}`,
-            collapsible: hasChildren ? 'header' : 'disabled',
-            extra: <PanelToolContainer
-                group={group}
-                opts={opts}
-                layerCount={group.getLayerCount()}
-                controller={controller}
-                allLayersOnMap={allLayersOnMap} />,
-            children: <LayerCollapsePanel key={group.getId()}
-                trimmed
-                selectedLayerIds={selectedLayerIds}
-                group={group}
-                openGroupTitles={openGroupTitles}
-                layerRows={layerRows}
-                opts={opts}
-                controller={controller}
-            />
-        };
-    });
-
+    const items = getCollapseItems(groups, openGroupTitles, selectedLayerIds, opts, controller);
     return (
         <StyledCollapse
             bordered
             activeKey={openGroupTitles}
             onChange={keys => controller.updateOpenGroupTitles(keys)}
-            items={groupItems}
+            items={items}
         />
     );
 };
